requirements.txt
aiogram==3.21
aiohttp
beautifulsoup4
python-dotenv